requirements.txt
yfinance
supabase
pandas
numpy
scikit-learn
joblib
python-dotenv